requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
googlesearch-python==1.2.3
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.7
Werkzeug==2.3.7
gunicorn==21.2.0